Added IWHttpRequest.
This wrapper API simplifies usage of NSURLConnection.
Performing POST or GET is as simple as passing the
string URL and a dictionary of parameters.

It can also accept callbacks for success, failure,
and a final callback.

It supports downloading of large files by setting
the destination path.

The API runs in full async, based on Apple's
recommendations when performing any network tasks.

The documentation for this API shall be done in the
README.

Signed-off-by: John Paul Alcala <jp@teamcodeflux.com>